refactor(main): main プロセス専用になった Config を src/main へ移動
renderer からの Config 直接参照が消えたため、「src/lib = renderer 系」の
区分に合わせて配置を修正する。これで src/lib に残るのは preload から使う
UI ユーティリティと ipcRenderer ラッパー・tRPC クライアントのみ。
